Updated Damage/OnHit to float from int
- Balancing Glacier Beam damage and ability slow (Crysta turns slower when casting beam)

+++ Assets/Scripts/Player/HealthPoints/NetworkPlayer_Health.cs
@@ -110,7 +110,7 @@ public class NetworkPlayer_Health : CharacterComponent, IHealthComponent
     }
 
     // Function only called on the server
-    public void OnTakeDamage(int damageAmount)
+    public void OnTakeDamage(float damageAmount)
     {
         if (damageAmount < 0)
         {
